Refactor tests and vaccine forms
