+++ src/__tests__/pages/VaccinePage.test.tsx
@@ -6,7 +6,7 @@ import { act } from "react-dom/test-utils";
 import { screen, waitFor } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 
-import { renderWithProviders } from "../utils/test-utils";
+import { renderWithProviders } from "../../utils/test-utils";
 
 import VaccinePage from "../../pages/VaccinPage";
 
@@ -31,7 +31,7 @@ const server = setupServer(
       ctx.json({
         vaccines: vaccinesMockData,
       }),
-      ctx.delay(150)
+      ctx.delay(50)
     );
   }),
 
@@ -64,7 +64,7 @@ const server = setupServer(
         },
         message: "Vaccine created successfully",
       }),
-      ctx.delay(150)
+      ctx.delay(50)
     );
   }),
 
@@ -98,7 +98,7 @@ const server = setupServer(
         },
         message: "Vaccine updated successfullyy",
       }),
-      ctx.delay(150)
+      ctx.delay(50)
     );
   }),
 
@@ -109,7 +109,7 @@ const server = setupServer(
       ctx.json({
         message: "Vaccine deleted successfully",
       }),
-      ctx.delay(150)
+      ctx.delay(50)
     );
   })
 );
@@ -160,7 +160,7 @@ describe("VaccineContent", () => {
           ctx.json({
             vaccines: [],
           }),
-          ctx.delay(150)
+          ctx.delay(50)
         );
       })
     );
