+++ src/utils/test-utils.tsx
@@ -4,8 +4,8 @@ import type { RenderOptions } from "@testing-library/react";
 import type { PreloadedState } from "@reduxjs/toolkit";
 import { Provider } from "react-redux";
 
-import type { AppStore, RootState } from "../../redux/store";
-import { setupStore } from "../../redux/store";
+import type { AppStore, RootState } from "../redux/store";
+import { setupStore } from "../redux/store";
 
 // This type interface extends the default options for render from RTL, as well
 // as allows the user to specify other things such as initialState, store.
